Rename rule S6855 to S4084
