+++ packages/jsts/src/rules/S4084/decorator.ts
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-// https://sonarsource.github.io/rspec/#/rspec/S6855/javascript
+// https://sonarsource.github.io/rspec/#/rspec/S4084/javascript
 
 import { Rule } from 'eslint';
 import { Node } from 'estree';
