+++ packages/jsts/src/rules/S4084/index.ts
